Conditional build dependent upon branch PR status.

+++ travis/cleanup.sh
@@ -1,3 +1,6 @@
 #!/usr/bin/env bash
 
-shred --remove ${TRAVIS_BUILD_DIR}/travis/encrypt-settings.xml
+if [ "${TRAVIS_BRANCH}" = 'dev' ] || [ "${TRAVIS_BRANCH}" = 'master' ]; then
+    shred --remove ${TRAVIS_BUILD_DIR}/travis/encrypt-settings.xml
+    shred --remove ${TRAVIS_BUILD_DIR}/travis/signing.asc
+fi
